mapshell: move shell logic to pipeline.shell module, adjust mapshell.py to use
